requirements.txt
streamlit
torch
torchvision
tensorflow
pillow
numpy
matplotlib
opencv-python
lime
scikit-image
